Update edit.blade.php
I'm not sure if there is a specific reason to use `@aware` here or not, but I noticed that it's usage is having an unintended side-effect in that the code is creating a `record` attribute on the anchor tag and the *whole* record is showing as the value of that attribute. As far as I can tell, this isn't needed and furthermore it could become a potential security issue in production environments.